+++ packages/sx.js/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @snapshot-labs/sx
 
+## 0.1.1
+
+### Patch Changes
+
+- 827b7b3: add OffchainEthereumSig client with basic vote support
+- ddcdc7f: use clientConfig.ethUrl instead of predefined RPC APIs for starknet strategies
+
 ## 0.1.0
 
 ### Patch Changes